Fix matching of Proc::check_args if cmdline contains non-exe path
In the current version Proc::check_args matches all arguments of the cmdline to
the needle to check if a program is already running. This check returns wrong
result if for example the path `/home/root/.local/share/remarkable/xochitl` is
somewhere in the cmdline arguments.

The new version just matches the first argument of the cmdline and only checks
for further args if the first program is a known interpreter (sh, bash, python).

Signed-off-by: Jan Luca Naumann <j.naumann@fu-berlin.de>