[remux] return to unthreaded way of launching applications
there is a bug where users can not kill xochitl through remux and
restart it. it shows up in various forms and seems to be related to
the way of launching the applications.

by using '&' (instead of running the program in just a separate thread),
this bug goes away.

(bug was introduced in 154af55c2d40a371e6d9447070976b937fc3a5eb)